Significant rearrangement of code
Am disentangling the logical components of the code in logic.js from the GUI components in gui.js and QED.html.  This was trickier than expected to pull off and required rethinking some of the architecture I had previously put in place.  Still need to update classes.txt to reflect the new structure.